add jupyter notebook for visualization
